AS#, Consolidate actions in projects folder -- Consolidated actions and reducers to the same place in projects.
