+++ src/projects/reducers/projectSearch.js
@@ -2,7 +2,7 @@ import {
   PROJECT_SEARCH_PENDING, PROJECT_SEARCH_SUCCESS, PROJECT_SEARCH_FAILURE,
   GET_PROJECTS_PENDING, GET_PROJECTS_SUCCESS, GET_PROJECTS_FAILURE,
   LOAD_MORE_PROJECTS, CLEAR_PROJECT_SEARCH, GET_PROJECTS_SEARCH_CRITERIA
-} from '../../../config/constants'
+} from '../../config/constants'
 
 export const initialState = {
   isLoading: true,
