+++ src/reducers/index.js
@@ -4,8 +4,8 @@ import loadUser from './loadUser'
 import { projectState } from '../projects/reducers/project'
 import { projectDashboard } from '../projects/reducers/projectDashboard'
 import navSearch from './navSearch'
-import projectSearch from '../projects/list/reducers/projectSearch'
-import projectSearchSuggestions from '../projects/list/reducers/projectSearchSuggestions'
+import projectSearch from '../projects/reducers/projectSearch'
+import projectSearchSuggestions from '../projects/reducers/projectSearchSuggestions'
 import members from './members'
 import alerts from './alerts'
